Move api_discovery to a separate package
