+++ pkg/utils/discovery/api_discover.go
@@ -1,4 +1,4 @@
-package utils
+package discovery
 
 import (
 	"fmt"
