+++ pkg/utils/discovery/api_discover_test.go
@@ -1,4 +1,4 @@
-package utils
+package discovery
 
 import "testing"
